+++ crates/wolf_engine_input/src/mouse.rs
@@ -1,0 +1,12 @@
+//! Provides types related to mouse input.
+
+/// Identifies a mouse button.
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub enum MouseButton {
+    Left,
+    Middle,
+    Right,
+    Forward,
+    Back,
+    Other(u32),
+}