+++ examples/input.rs
@@ -50,5 +50,10 @@ fn process_input(input: &Input) {
             scancode,
             keycode,
         } => println!("Raw Key: {state:?}, {scancode:?}, {keycode:?}"),
+        Input::MouseMoved { x, y } => println!("Mouse Moved: {x}, {y}"),
+        Input::RawMouseMoved { delta_x, delta_y } => {
+            println!("Raw Mouse Moved: {delta_x}, {delta_y}")
+        }
+        Input::MouseButton { state, button } => println!("Mouse Button: {button:?} {state:?}"),
     }
 }